+++ ending-credits-api/src/main/java/com/hanaro/endingcredits/endingcreditsapi/domain/product/dto/RetirementPensionDetailResponseDto.java
@@ -1,0 +1,4 @@
+package com.hanaro.endingcredits.endingcreditsapi.domain.product.dto;
+
+public class RetirementPensionDetailResponseDto {
+}
